refactor countdown timer
