+++ src/utils/azorius.ts
@@ -103,7 +103,7 @@ export const mapProposalCreatedEventToProposal = async (
     logError('Error while getting strategy quorum');
     quorum = BigNumber.from(0);
   }
-  const deadline = await getTimeStamp(endBlock, provider);
+  const deadlineSeconds = await getTimeStamp(endBlock, provider);
   const state = await getAzoriusProposalState(
     strategyContract,
     azoriusContract.asSigner,
@@ -141,7 +141,7 @@ export const mapProposalCreatedEventToProposal = async (
     proposer,
     startBlock: BigNumber.from(startBlock),
     transactionHash,
-    deadline: deadline,
+    deadlineMs: deadlineSeconds * 1000,
     state,
     govTokenAddress: await strategyContract.governanceToken(),
     votes,
